Cleaned up code, renamed IO to OperatorInputs, changed TeleopCommand call
