+++ src/main/java/frc/robot/Robot.java
@@ -19,8 +19,8 @@ import frc.robot.subsystems.drive.Gyroscope;
 import frc.robot.commands.AutonomousCommand;
 import frc.robot.commands.LightsCommand;
 import frc.robot.commands.TeleopCommand;
-import frc.robot.io.Controls;
-import frc.robot.io.IO;
+import frc.robot.operatorInputs.Controls;
+import frc.robot.operatorInputs.OperatorInputs;
 
 public class Robot extends TimedRobot {
 
@@ -36,13 +36,13 @@ public class Robot extends TimedRobot {
   private Limelight limelight;
   private Vision vision;
   private Gyroscope gyro;
-  private IO io;
+  private OperatorInputs operatorInputs;
   private AutonomousCommand autonomousCommand;
+  private TeleopCommand teleopCommand;
   private LightsCommand lightsCommand;
   private ShooterSpeedController shooterSpeedController;
   private HopperController hopperController;
   private IndexerController indexerController;
-
 
   // Constants
   private final int JOYSTICK_PORT = 1;
@@ -71,16 +71,17 @@ public class Robot extends TimedRobot {
     this.controls = new Controls(new Joystick(JOYSTICK_PORT));
     this.shooter = new Shooter();
     this.lights = new Lights(9, 60, 50);
-    this.vision = new Vision(limelight,lights);
+    this.vision = new Vision(limelight, lights);
     this.limelight = new Limelight();
     this.lightsCommand = new LightsCommand(lights);
     this.shooterSpeedController = new ShooterSpeedController();
     this.hopperController = new HopperController();
     this.indexerController = new IndexerController();
-    this.io = new IO(controls, drive, gyro, shooter, shooterSpeedController, hopperController, indexerController,limelight);
-
+    this.operatorInputs = new OperatorInputs(controls, drive, gyro, shooter, shooterSpeedController, hopperController,
+        indexerController, limelight);
     this.lightsCommand = new LightsCommand(this.lights);
     this.autonomousCommand = new AutonomousCommand(distance, maxSpeed, drive, pValue);
+    this.teleopCommand = new TeleopCommand(this.controls, this.drive);
   }
 
   /**
@@ -115,14 +116,14 @@ public class Robot extends TimedRobot {
   public void autonomousInit() {
     m_autoSelected = m_chooser.getSelected();
     switch (m_autoSelected) {
-    case kCustomAuto:
-      break;
-    case kDefaultAuto:
-    default:
-      if (autonomousCommand != null) {
-        autonomousCommand.schedule();
-      }
-      break;
+      case kCustomAuto:
+        break;
+      case kDefaultAuto:
+      default:
+        if (autonomousCommand != null) {
+          autonomousCommand.schedule();
+        }
+        break;
     }
   }
 
@@ -133,7 +134,9 @@ public class Robot extends TimedRobot {
 
   @Override
   public void teleopInit() {
-    drive.setDefaultCommand(new TeleopCommand(controls, drive));
+    if (teleopCommand != null) {
+      teleopCommand.schedule();
+    }
   }
 
   // use this to override the algorithm and just use a speed
@@ -142,14 +145,15 @@ public class Robot extends TimedRobot {
     this.overrideSpeed = overrideSpeed;
     shooterSpeedController.setLaunchSpeed(this.overrideSpeed);
   }
- 
+
   @Override
   public void teleopPeriodic() {
     CommandScheduler.getInstance().run();
-      if (controls.getLeftTrigger()) {
-        indexerController.outtake();
-      }
+    if (controls.getLeftTrigger()) {
+      indexerController.outtake();
+    }
   }
+
   /**
    * This function is called periodically during test mode.
    */
