+++ src/main/java/frc/robot/operatorInputs/Controls.java
@@ -1,4 +1,4 @@
-package frc.robot.io;
+package frc.robot.operatorInputs;
 
 import edu.wpi.first.wpilibj.Joystick;
 import edu.wpi.first.wpilibj2.command.button.JoystickButton;
