+++ src/main/java/frc/robot/operatorInputs/OperatorInputs.java
@@ -1,4 +1,4 @@
-package frc.robot.io;
+package frc.robot.operatorInputs;
 
 import frc.robot.commands.GyroTurnCommand;
 import frc.robot.commands.ReorientToFieldCommand;
@@ -12,9 +12,9 @@ import frc.robot.subsystems.shooter.Shooter;
 import frc.robot.subsystems.shooter.ShooterSpeedController;
 import frc.robot.subsystems.vision.Limelight;
 
-public class IO {
+public class OperatorInputs {
 
-  public IO(Controls controls, Drive drive, Gyroscope gyro, Shooter shooter, ShooterSpeedController shooterSpeedController,
+  public OperatorInputs(Controls controls, Drive drive, Gyroscope gyro, Shooter shooter, ShooterSpeedController shooterSpeedController,
       HopperController hopperController, IndexerController indexController,Limelight limelight) {
     controls.getJoystickXButton().whenPressed(new GyroTurnCommand(gyro, drive, 180));
     controls.getRightJoystickBumper().whenPressed(new GyroTurnCommand(gyro, drive, -90));
